+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## General updates 
 
+* Refactored wasting case definers to halt execution and raise an error when oedema contains missing values.
 * Revised the mathematical expression for the age-weighting approach in `mw_estimate_age_weighted_prev_muac()`. 
 * Added a hyperlink to the internet‑deployed `mwana` Shiny App (`mwanaApp`) in the website’s list of links.